feat: set camera (constant speed)

+++ astr/windows_api_lecture/Client/CAnimation.h
@@ -31,7 +31,7 @@ public:
 		m_fAccTime = 0.f;
 	}
 	tAnimFrm& GetFrame(int _iIdx) { return m_vecFrm[_iIdx]; }
-	int GetMaxFrame() { return m_vecFrm.size(); }
+	UINT GetMaxFrame() { return (UINT)m_vecFrm.size(); }
 private:
 	void SetName(const wstring& _strName) { m_strName = _strName; }
 public:
